Filtering by incrementation corrected.

+++ app/src/main/java/com/eltescode/modules/features/modules/presentation/main_screen/ModulesView.kt
@@ -118,7 +118,7 @@ fun ModulesView(
                                 .fillMaxWidth(),
                             modules = state.modules
                                 .filter { module -> module.epochDay == epochDate }
-                                .sortedByDescending { it.incrementation },
+                                .sortedByDescending { it.incrementation.toInt() },
                             onAddButtonClick = {
 
                                 onEvent(MainScreenEvents.OnAddButtonClick(epochDate))
